# M2-Challenge-Yusuf-Hassan/MonthAndMathService/src/main/java/com/company/MonthAndMathService/model/CustomErrorResponse.java
package com.company.MonthAndMathService.model;

import java.time.LocalDateTime;

public class CustomErrorResponse {

    private String errorMsg;
    private int status;
    private String errorCode;
    private LocalDateTime timestamp;

    public CustomErrorResponse(String errorMsg, String errorCode) {
        this.errorMsg = errorMsg;
        this.errorCode = errorCode;
    }

    public String getErrorMsg() {
        return errorMsg;
    }

    public void setErrorMsg(String errorMsg) {
        this.errorMsg = errorMsg;
    }

    public int getStatus() {
        return status;
    }

    public void setStatus(int status) {
        this.status = status;
    }

    public String getErrorCode() {
        return errorCode;
    }

    public void setErrorCode(String errorCode) {
        this.errorCode = errorCode;
    }

    public LocalDateTime getTimestamp() {
        return timestamp;
    }

    public void setTimestamp(LocalDateTime timestamp) {
        this.timestamp = timestamp;
    }
}