# M2-Challenge-Yusuf-Hassan/MonthAndMathService/src/main/java/com/company/MonthAndMathService/model/MathSolution.java
package com.company.MonthAndMathService.model;

import org.springframework.validation.annotation.Validated;

import javax.validation.constraints.*;
import java.util.Objects;

public class MathSolution {

    @NotNull(message = "Please enter a valid number for operand 1 - field should not be blank.")
    private Integer operand1;
    @NotNull(message = "Please enter a valid number for operand 2 - field should not be blank.")
    private Integer operand2;
    private String operation;
    private int answer;

    public MathSolution(Integer operand1, Integer operand2) {
        this.operand1 = operand1;
        this.operand2 = operand2;
    }

    public MathSolution() {
    }

    public int getOperand1() {
        return operand1;
    }

    public void setOperand1(int operand1) {
        this.operand1 = operand1;
    }

    public int getOperand2() {
        return operand2;
    }

    public void setOperand2(int operand2) {
        this.operand2 = operand2;
    }

    public String getOperation() {
        return operation;
    }

    public void setOperation(String operation) {
        this.operation = operation;
    }

    public int getAnswer() {
        return answer;
    }

    public void setAnswer(int answer) {
        this.answer = answer;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        MathSolution that = (MathSolution) o;
        return operand1 == that.operand1 && operand2 == that.operand2 && answer == that.answer && Objects.equals(operation, that.operation);
    }

    @Override
    public int hashCode() {
        return Objects.hash(operand1, operand2, operation, answer);
    }
}